document_manager/pyproject.toml
[tool.poetry]
name = "document-manager"
version = "0.1.0"
description = "Real World Software Development Exercise"
authors = ["Andres De Las Salas <andresdlassalas@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
coverage = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
